WIP filters component for observations and documents db
