+++ css/components/ui/_filters.scss
@@ -1,21 +1,32 @@
-
 .c-filters {
-  background-color: $color-secondary;
-  padding: 30px 30px 65px;
+  padding-top: 36px;
 
   .c-title {
     margin-top: 0;
+    font-size: 36px;
+    letter-spacing: 0;
+    line-height: 41px;
   }
 
-  .field {
-    .title {
-      font-family: $font-family-georgia;
-      color: $color-white;
-      text-transform: capitalize;
-      margin-bottom: 2px;
-      font-size: $font-size-big;
-      line-height: 26px;
-      font-weight: normal;
+  .filters-wrapper {
+    background-color: $color-primary;
+    padding: 30px 30px 65px;
+
+    .flex-wrapper {
+      display: flex;
+      flex-flow: row wrap;
+
+      .field {
+        .title {
+          font-family: $font-family-georgia;
+          color: $color-white;
+          text-transform: capitalize;
+          margin-bottom: 2px;
+          font-size: $font-size-big;
+          line-height: 26px;
+          font-weight: normal;
+        }
+      }
     }
   }
 }
